database/requirements.txt
psycopg2-binary
python-dotenv==1.0.0
requests==2.31.0